+++ CHANGELOG.md
@@ -24,7 +24,9 @@ should change the heading of the (upcoming) version to include a major version b
 - Make label generation consistent with other themes by refactoring the code into the `FieldTemplate` instead of having the widgets implementing the label, fixing [#2007](https://github.com/rjsf-team/react-jsonschema-form/issues/2007)
 
 ## @rjsf/chakra-ui
-- Added support for `chakra-react-select` v4, fixing [#3152](https://github.com/rjsf-team/react-jsonschema-form/issues/3152).
+- Added support for `chakra-react-select` v4, fixing [#3152](https://github.com/rjsf-team/react-jsonschema-form/issues/3152)
+- In `SelectWidget` use `Select` from `chakra-react-select` for both single- and multiple-choice select
+- In `SelectWidget` multiple-choice select display label rather than value for selected items
 
 ## @rjsf/core
 - Extended `Form.onChange` to optionally return the `id` of the field that caused the change, fixing [#2768](https://github.com/rjsf-team/react-jsonschema-form/issues/2768)